code/aisentimentechnoverse/requirements.txt
altair==5.0.1
asgiref==3.7.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dj-database-url==2.0.0
Django==4.2.4
django-heroku==0.3.1
djangorestframework==3.14.0
gitdb==4.0.10
GitPython==3.1.32
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.25.2
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
plotly==5.16.0
protobuf==4.24.0
psycopg2==2.9.7
pyarrow==12.0.1
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
smmap==5.0.0
sqlparse==0.4.4
streamlit==1.25.0
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.21.2
whitenoise==6.5.0
zipp==3.16.2
